+++ gui/configuration_tab.py
@@ -1,3 +1,6 @@
+import sys
+sys.path.append("..")
+
 from javax.swing import DefaultComboBoxModel
 from java.awt.event import ActionListener
 from javax.swing import JToggleButton
@@ -10,7 +13,7 @@ from javax.swing import JCheckBox
 from javax.swing import JButton
 from javax.swing import JPanel
 
-from table import clearList
+from table.table import clearList
 
 class ConfigurationTab():
     def __init__(self, extender):
